Do not treat --trusted-host as a requirement
Prior to this commit, requirement files with a --trusted-host argument
in them were being treated as a requirement and pip-api would raise an
error due to the URL not containing an egg specifier.

This directly affect's pip-audit's ability to audit such requirement
files.

The commit fixes the issue by adding --trusted-host to the list of
known arguments that can be ignored for not being a requirement.